override feature for test-data

+++ src/main/java/com/polarbookshop/catalogservice/demo/BookDataLoader.java
@@ -1,5 +1,6 @@
 package com.polarbookshop.catalogservice.demo;
 
+import com.polarbookshop.catalogservice.config.properties.PolarProperties;
 import com.polarbookshop.catalogservice.domain.Book;
 import com.polarbookshop.catalogservice.domain.BookRepository;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
@@ -15,14 +16,24 @@ import java.util.stream.Stream;
  */
 public class BookDataLoader {
     private final BookRepository bookRepository;
+    private final PolarProperties polarProperties;
 
-    public BookDataLoader(BookRepository bookRepository) {
+    public BookDataLoader(BookRepository bookRepository, PolarProperties polarProperties) {
         this.bookRepository = bookRepository;
+        this.polarProperties = polarProperties;
     }
 
     @EventListener(ApplicationReadyEvent.class)
     public void loadBookTestData() {
-        bookRepository.deleteAll();
+        if (polarProperties.testData().isEnable()) {
+            if (bookRepository.count() == 0 || polarProperties.testData().doOverride()) {
+                bookRepository.deleteAll();
+                generateData();
+            }
+        }
+    }
+
+    private void generateData() {
         bookRepository.saveAll(Stream.of(
             Book.of("1234567891", "Northern Lights", "Lyra Silverstar", 9.90, "TMC Production"),
             Book.of("1234567892", "Polar Journey", "Lorek Polarson", 12.90, "BNY Mellon")
